# #53655: Improve speed of DOMNode::C14N() on large XML documents